Separate Customer and Merchant routes into nested routes (#134)
* Nest customer and merchant routes

* Use destructuring to call match.path in Routes.tsx

* Remove unnecessary Customer Pages comment
